Make runners for each runnable entity
